chore(dev): Add support for yamllint to lint conference data
Conference data is persisted as .yaml files. It is mission-critical that all .yaml files are valid. To address this concern, yamllint is added to CI and to DevContainer. This allows YAML linting during CI runs (also for new PRs) and locally when running in a DevContainer or when yamllint is installed on the host.

fixes #6

Signed-off-by: Thorsten Hans <thorsten.hans@gmail.com>